WIP: Remove finalizer from cluster if AKO is not present
AKOO adds a finalizer to the cluster , in case AKO needs to clean up resources in AVI
Controller before cluster is deleted.  But in some cases, AKO might have crashed or
isn't even there, for example, when cluster is not created successfully.
In that case, AKOO will not remove the finalizer since AKO can't report the
status correctly.

If the AKO packge is not installed on the cluster and the cluster is marked
for deletion, AKOO will remove the finalizers from the cluster